Intrinsic: support intrinsic for RdModel (mostly for sake of UProperty)
